+++ src/rs-core/bindings/js_functions/mod.rs
@@ -519,6 +519,7 @@ pub fn jsUpdateContentInfo(
     minimum_position: Option<f64>,
     maximum_position: Option<f64>,
     playlist_nat: PlaylistNature,
+    uses_program_date_time: bool,
 ) {
     unsafe {
         __js_func__update_content_info(
@@ -527,6 +528,7 @@ pub fn jsUpdateContentInfo(
             bool_to_raw(maximum_position.is_some()),
             maximum_position.unwrap_or(0.),
             playlist_nat as u32,
+            bool_to_raw(uses_program_date_time),
         )
     }
 }